+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # HEAD
 
+- Include calibration configuration and data files in the database ([#18](https://github.com/ziotom78/qutedb/pull/18))
+
 # 0.4.1
 
 - Be more tolerant when multiple files match a mask ([#17](https://github.com/ziotom78/qutedb/pull/17))